Introduce UniversalPicker component
