+++ idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/common/picker/PickerEvent.kt
@@ -1,0 +1,9 @@
+package io.github.composegears.valkyrie.ui.common.picker
+
+import java.nio.file.Path
+
+sealed interface PickerEvent {
+    data class PickDirectory(val path: Path) : PickerEvent
+    data class PickFiles(val paths: List<Path>) : PickerEvent
+    data class ClipboardText(val text: String) : PickerEvent
+}
