+++ idea-plugin/src/main/resources/messages/Valkyrie.properties
@@ -6,3 +6,9 @@ intro.card.iconpack.title=IconPack
 intro.card.iconpack.description=Create organized icon pack with batch export into your project
 settings.export.composecolor=Use Compose colors
 settings.export.composecolor.description=Predefined Compose colors will be used instead of hex color codes (e.g. Color.Black instead of Color(0xFF000000))
+picker.dnd=Drag \\& drop
+picker.clipboard.mac=Cmd+V to paste from clipboard
+picker.clipboard.other=Ctrl+V to paste from clipboard
+picker.dnd.or=or
+picker.pick.directory=Pick dir
+picker.pick.files=Pick files
